// src/app/enum/region-enum.ts
export enum RegionEnum {
  REGION_ARA = 'Auvergne-Rhône-Alpes',
  REGION_BFC = 'Bourgogne-Franche-Comté',
  REGION_BRE = 'Bretagne',
  REGION_CVL = 'Centre-Val de Loire',
  REGION_20R = 'Corse',
  REGION_GES = 'Grand Est',
  REGION_HDF = 'Hauts-de-France',
  REGION_IDF = 'Île-de-France',
  REGION_NOR = 'Normandie',
  REGION_NAQ = 'Nouvelle-Aquitaine',
  REGION_OCC = 'Occitanie',
  REGION_PDL = 'Pays de la Loire',
  REGION_PAC = "Provence-Alpes-Côte d'Azur",
  REGION_GLP = 'Guadeloupe',
  REGION_GUF = 'Guyane',
  REGION_MTQ = 'Martinique',
  REGION_REU = 'La Réunion',
  REGION_MYT = 'Mayotte',
}